// src/integrations/supabase/firestore/schema.generated.ts
export type TableName = keyof Database["public"]["Tables"] & string;

export type ForeignKey = {
  readonly columns: readonly string[];
  readonly isOneToOne: boolean;
  readonly referencedRelation: TableName;
  readonly referencedColumns: readonly string[];
};